redo inventory apis and typings
