+++ src/core/inventory-actions.ts
@@ -1,16 +1,36 @@
-import { InventoryAction, InventoryLayerKind } from '../core/inventory-layers'
+import { InventoryLayerKind } from '../core/inventory-layers'
+
+export interface InventoryAction {
+  /** Which layer this action belongs to */
+  layer: InventoryLayerKind
+  /** Key binding to activate this action */
+  keybind: string
+  /** Action name. Will be formatted and appended to layer name. Final command value would be :Veonim ${layer}-${command}*/ 
+  name: string
+  /** User friendly description provided in the UI */
+  description: string
+  /** Callback will be executed when this action is selected */
+  onAction: () => any
+  /** Indicate to the user that this action is experimental. Default: FALSE */
+  experimental?: boolean
+  /** Hide this action from the inventory menu. Otherwise will show up in the inventory search menu. Default: FALSE */ 
+  hidden?: boolean
+}
 
 const mod = (modulePath: string, func = 'default') => {
   return require(`../${modulePath}`)[func]
 }
 
+// TODO: allow actions to be registered as 'hidden'. these will not be displayed
+// in the UI as options, but can be found in the fuzzy search menu. useful for
+// some less common actions
 const actions: InventoryAction[] = [
   {
     layer: InventoryLayerKind.Project,
     keybind: 'f',
     name: 'Files',
     description: 'Find files in project',
-    onAction: mod('components/files'),
+    onAction: mod('components/filesz'),
   },
   {
     layer: InventoryLayerKind.Project,
@@ -43,6 +63,10 @@ const actions: InventoryAction[] = [
 ]
 
 // TODO: register all these actions as neovim commands
+
+const wut = actions[0]
+console.log('wut', wut)
+wut.onAction()
 
 export default {
   list: () => actions,
