+++ src/neovim/api.ts
@@ -1,8 +1,8 @@
 import { VimMode, VimOption, BufferEvent, HyperspaceCoordinates, BufferType, BufferHide, BufferOption, Color, Buffer, Window, Tabpage, GenericCallback, Keymap } from '../neovim/types'
 import { Api, ExtContainer, Prefixes, Buffer as IBuffer, Window as IWindow, Tabpage as ITabpage } from '../neovim/protocol'
 import { asColor, is, onFnCall, onProp, prefixWith, uuid, Watcher, GenericEvent } from '../support/utils'
+import { InventoryAction } from '../core/inventory-actions'
 import { normalizeVimMode } from '../support/neovim-utils'
-import { InventoryAction } from '../core/inventory-layers'
 import { SHADOW_BUFFER_TYPE } from '../support/constants'
 import { Autocmd, Autocmds } from '../core/vim-startup'
 import { watchConfig } from '../config/config-reader'
